refactor nostr client
