+++ app/src/main/java/com/koalasat/pokey/ui/relays/RelaysFragment.kt
@@ -20,14 +20,8 @@ import com.koalasat.pokey.R
 import com.koalasat.pokey.database.AppDatabase
 import com.koalasat.pokey.database.RelayEntity
 import com.koalasat.pokey.databinding.FragmentRelaysBinding
-import com.koalasat.pokey.models.ExternalSigner
-import com.vitorpamplona.ammolite.relays.COMMON_FEED_TYPES
-import com.vitorpamplona.ammolite.relays.Client
-import com.vitorpamplona.ammolite.relays.Relay
-import com.vitorpamplona.ammolite.relays.RelayPool
-import com.vitorpamplona.quartz.encoders.toHexKey
-import com.vitorpamplona.quartz.events.Event
-import com.vitorpamplona.quartz.utils.TimeUtils
+import com.koalasat.pokey.models.NostrClient
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -88,7 +82,9 @@ class RelaysFragment : Fragment() {
             }
         }
         binding.publishPublicRelay.setOnClickListener {
-            publishPublicRelays()
+            CoroutineScope(Dispatchers.IO).launch {
+                context?.let { it1 -> NostrClient.publishPublicRelays(it1) }
+            }
         }
         binding.reloadPublicRelays.setOnClickListener {
             Pokey.updateLoadingPublicRelays(true)
@@ -127,7 +123,9 @@ class RelaysFragment : Fragment() {
             }
         }
         binding.publishPrivateRelay.setOnClickListener {
-            publishPrivateRelays()
+            CoroutineScope(Dispatchers.IO).launch {
+                context?.let { it1 -> NostrClient.publishPrivateRelays(it1) }
+            }
         }
         binding.reloadPrivateRelay.setOnClickListener {
             Pokey.updateLoadingPrivateRelays(true)
@@ -160,28 +158,8 @@ class RelaysFragment : Fragment() {
 
     private fun insertRelay(url: String, kind: Int) {
         lifecycleScope.launch {
-            val hexKey = Pokey.getInstance().getHexKey()
-            val dao = context?.let { AppDatabase.getDatabase(it, hexKey).applicationDao() }
-            if (dao != null && url.isNotEmpty()) {
-                val existsRelay = withContext(Dispatchers.IO) { dao.existsRelay(url, kind) }
-                if (existsRelay < 1) {
-                    val entity = RelayEntity(id = 0, url, kind = kind, createdAt = 0, read = 1, write = 1)
-                    withContext(Dispatchers.IO) { dao.insertRelay(entity) }
-
-                    val relay = RelayPool.getRelay(url)
-                    if (Pokey.isEnabled.value == true && relay == null) {
-                        RelayPool.addRelay(
-                            Relay(
-                                entity.url,
-                                read = entity.read == 1,
-                                write = entity.write == 1,
-                                forceProxy = false,
-                                activeTypes = COMMON_FEED_TYPES,
-                            ),
-                        )
-                        RelayPool.connectAndSendFiltersIfDisconnected()
-                    }
-                }
+            CoroutineScope(Dispatchers.IO).launch {
+                context?.let { NostrClient.addRelay(it, url, kind) }
                 loadRelays()
             }
         }
@@ -189,12 +167,8 @@ class RelaysFragment : Fragment() {
 
     private fun reconnectRelays(kind: Int) {
         lifecycleScope.launch {
-            val hexKey = Pokey.getInstance().getHexKey()
-            val dao = context?.let { AppDatabase.getDatabase(it, hexKey).applicationDao() }
-            if (dao != null) {
-                Pokey.getInstance().stopService()
-                withContext(Dispatchers.IO) { dao.deleteRelaysByKind(kind) }
-                Pokey.getInstance().startService()
+            CoroutineScope(Dispatchers.IO).launch {
+                context?.let { NostrClient.reconnectInbox(it, kind) }
             }
         }
     }
@@ -240,78 +214,6 @@ class RelaysFragment : Fragment() {
                     confirmation()
                 }
             }
-        }
-    }
-
-    private fun publishPrivateRelays() {
-        val pubKey = Pokey.getInstance().getHexKey()
-        val createdAt = TimeUtils.now()
-        val kind = 10050
-        val content = ""
-        val tags = privateList.map { arrayOf("relay", it.url) }.toTypedArray()
-        val id = Event.generateId(pubKey, createdAt, kind, tags, content).toHexKey()
-        val event =
-            Event(
-                id = id,
-                pubKey = pubKey,
-                createdAt = createdAt,
-                kind = kind,
-                tags = tags,
-                content = content,
-                sig = "",
-            )
-        ExternalSigner.sign(event) {
-            val signeEvent = Event(
-                id = id,
-                pubKey = pubKey,
-                createdAt = createdAt,
-                kind = kind,
-                tags = tags,
-                content = content,
-                sig = it,
-            )
-            Log.d("Pokey", "Relay private list : ${signeEvent.toJson()}")
-            Client.send(signeEvent)
-        }
-    }
-
-    private fun publishPublicRelays() {
-        val pubKey = Pokey.getInstance().getHexKey()
-        val createdAt = TimeUtils.now()
-        val kind = 10002
-        val content = ""
-        val tags = publicList.map {
-            var tag = arrayOf("r", it.url)
-            if (it.read == 1 && it.write == 0) {
-                tag += "read"
-            } else if (it.read == 0 && it.write == 1) {
-                tag += "write"
-            }
-            tag
-        }.toTypedArray()
-        val id = Event.generateId(pubKey, createdAt, kind, tags, content).toHexKey()
-        val event =
-            Event(
-                id = id,
-                pubKey = pubKey,
-                createdAt = createdAt,
-                kind = kind,
-                tags = tags,
-                content = content,
-                sig = "",
-            )
-        ExternalSigner.sign(event) {
-            val signeEvent = Event(
-                id = id,
-                pubKey = pubKey,
-                createdAt = createdAt,
-                kind = kind,
-                tags = tags,
-                content = content,
-                sig = it,
-            )
-            Log.d("Pokey", "Relay public list : ${signeEvent.toJson()}")
-            Client.send(signeEvent)
         }
     }
 }
